API-LLM/requirements.txt
fastapi
ollama
uvicorn
python-dotenv
requests
